backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
xlwings==0.30.13
scikit-learn==1.3.2
prophet==1.1.5
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
